Moved classes from com.hivemq.extensions.services.auth to com.hivemq.extensions.auth
