+++ src/main/java/com/hivemq/extensions/auth/ReAuthTask.java
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-package com.hivemq.extensions.services.auth;
+package com.hivemq.extensions.auth;
 
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import com.hivemq.extension.sdk.api.auth.EnhancedAuthenticator;
 import com.hivemq.extension.sdk.api.auth.parameter.AuthenticatorProviderInput;
 import com.hivemq.extension.sdk.api.packets.auth.AuthReasonCode;
 import com.hivemq.extensions.client.ClientAuthenticators;
+import com.hivemq.extensions.services.auth.WrappedAuthenticatorProvider;
 
 /**
  * @author Daniel Krüger
