+++ src/test/java/com/hivemq/extensions/auth/parameter/ModifiableClientSettingsImplTest.java
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.extensions.services.auth;
+package com.hivemq.extensions.auth.parameter;
 
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import com.hivemq.extension.sdk.api.auth.parameter.OverloadProtectionThrottlingLevel;
+import com.hivemq.extensions.auth.parameter.ModifiableClientSettingsImpl;
 import org.junit.Before;
 import org.junit.Test;
